ai-realtime-assistant-texml/requirements.txt
aider==0.2.6
aider-chat==0.59.1
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
argcomplete==3.5.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.0
autoflake==2.3.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.3.0
bleach==6.1.0
blinker==1.7.0
boto3==1.34.72
botocore==1.34.77
bs4==0.0.2
build==1.2.1
CacheControl==0.14.0
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
configparser==7.0.0
crashtest==0.4.1
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
discord==2.3.2
discord.py==2.3.2
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
docutils==0.20.1
dulwich==0.21.7
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.16.1
flake8==7.1.1
Flask==3.0.2
Flask-Session==0.8.0
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.3
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.1.0
kinde-python-sdk==1.2.3
litellm==1.47.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
maxminddb==2.6.0
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
networkx==3.2.1
nh3==0.2.17
notion-client==2.2.1
numpy==1.26.4
openai==1.47.0
openapi-client==1.1.7
openapi-python-client==0.19.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-review==1.3.0
pipenv==2024.0.1
pipfile==0.0.2
pipreqs==0.5.0
pipx==1.6.0
pkginfo==1.10.0
platformdirs==4.2.0
playwright==1.45.0
PlexAPI==4.15.11
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt_toolkit==3.0.47
proxyhub==0.0.1a5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==11.1.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pypager==3.0.1
pypandoc==1.13
pyperclip==1.9.0
pyproject-api==1.6.1
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
pyzmq==25.1.2
rapidfuzz==3.7.0
readme_renderer==43.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.0
rfc3986==2.0.0
rich==13.8.1
rpds-py==0.20.0
ruff==0.3.5
s3transfer==0.10.1
safetensors==0.4.2
scipy==1.13.1
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
-e git+https://github.com/team-telnyx/telnyx-python@6c1b15032389a620b9796cf7f093e8fcae7a4d91#egg=telnyx
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.4
torch==2.2.2
tornado==6.4
tox==4.14.2
tqdm==4.66.5
traitlets==5.14.2
transformers==4.39.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trove-classifiers==2024.3.25
twine==5.0.0
typer==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
userpath==1.9.2
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.2
xattr==1.1.0
yarg==0.1.9
yarl==1.11.1
zipp==3.20.2
